+++ src/components/TaskDialog.tsx
@@ -4,10 +4,10 @@ import {
     DialogContent,
     DialogTitle,
     Button,
+    CircularProgress,
 } from '@mui/material'
-import TaskFormProvider, { updateFunctionState } from './TaskFormProvider'
-import { useRecoilState } from 'recoil'
-import { formIsValidState } from './TaskFormProvider'
+import { useTaskForm } from '../hooks/UseTaskForm'
+import TaskFormDisplay from './TaskForm'
 
 interface TaskDialogProps {
     open: boolean
@@ -20,14 +20,27 @@ export const TaskDialog: React.FC<TaskDialogProps> = ({
     handleClose,
     selectedId,
 }) => {
-    //更新処理
-    const [updateFunction] = useRecoilState(updateFunctionState)
-    const [formIsValid] = useRecoilState(formIsValidState)
+    const {
+        register,
+        onSubmit,
+        data,
+        toggleStatus,
+        status,
+        formState,
+        isLoading,
+        setValue,
+    } = useTaskForm(Number(selectedId), handleClose, open)
 
-    const handleUpdateClick = () => {
-        if (updateFunction && formIsValid) {
-            updateFunction()
-        }
+    if (!open) {
+        return null
+    }
+
+    if (isLoading || status === undefined) {
+        return <CircularProgress />
+    }
+
+    const handleCloseResetStatus = () => {
+        handleClose()
     }
 
     return (
@@ -42,17 +55,22 @@ export const TaskDialog: React.FC<TaskDialogProps> = ({
         >
             <DialogTitle>タスク詳細</DialogTitle>
             <DialogContent>
-                <TaskFormProvider
+                <TaskFormDisplay
                     mode={1}
-                    id={Number(selectedId)}
-                    handleClose={handleClose}
+                    register={register}
+                    onSubmit={onSubmit}
+                    data={data}
+                    toggleStatus={toggleStatus}
+                    status={status}
+                    formState={formState}
+                    setValue={setValue}
                 />
             </DialogContent>
             <DialogActions>
-                <Button onClick={handleUpdateClick} color="primary">
+                <Button onClick={onSubmit} color="primary">
                     更新
                 </Button>
-                <Button onClick={handleClose} color="primary">
+                <Button onClick={handleCloseResetStatus} color="primary">
                     閉じる
                 </Button>
             </DialogActions>
